requirement.txt
matplotlib==3.5.1
numpy==1.24.3
rasterio==1.2.10
scikit-learn==1.2.2
scipy==1.10.1
torchvision==0.12.0
tqdm==4.65.0
yacs==0.1.6
yaml==0.2.5
